End of Semester 1. Status Report handed in.
Status report, with most importantly a detailed plan of next steps
in the project.
Notebook with working training loop, cleaned data and 1D
convolutions.
Plan is updated to include information from the status report. This
is open to change and at the start of each week more detail could
be added to the weeks plan.